Embark on a journey to create a rewrite tool.
