+++ mesonbuild/interpreterbase.py
@@ -1,0 +1,54 @@
+# Copyright 2016 The Meson development team
+
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+
+#     http://www.apache.org/licenses/LICENSE-2.0
+
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+# This class contains the basic functionality needed to run any interpreter
+# or an interpreter-based tool.
+
+from . import mparser
+from . import mesonlib
+from . import environment
+
+import os
+
+class InterpreterException(mesonlib.MesonException):
+    pass
+
+class InvalidCode(InterpreterException):
+    pass
+
+class InvalidArguments(InterpreterException):
+    pass
+
+
+
+class InterpreterBase:
+    def __init__(self, source_root, subdir):
+        self.source_root = source_root
+        self.funcs = {}
+        self.subdir = subdir
+
+    def load_root_meson_file(self):
+        mesonfile = os.path.join(self.source_root, self.subdir, environment.build_filename)
+        if not os.path.isfile(mesonfile):
+            raise InvalidArguments('Missing Meson file in %s' % mesonfile)
+        with open(mesonfile, encoding='utf8') as mf:
+            code = mf.read()
+        if len(code.strip()) == 0:
+            raise InvalidCode('Builder file is empty.')
+        assert(isinstance(code, str))
+        try:
+            self.ast = mparser.Parser(code).parse()
+        except mesonlib.MesonException as me:
+            me.file = environment.build_filename
+            raise me
